% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_matrix.R
\name{K_matrix}
\alias{K_matrix}
\title{Generate the structure/penalty matrix for sarim formulae}
\usage{
K_matrix(dimension, penalty, data)
}
\arguments{
\item{dimension}{Generate a squared matrix by (dimension x dimension).}

\item{penalty}{"identity", "rw1" or "rw2" for structure/penalty matrix.}

\item{data}{only interesting if penalty "gmrf" is choosen, in this case the data is returned.}
}
\description{
Function used in definition of smooth terms within sarim model formulae.
}
