\name{Sarim-package}
\alias{Sarim-package}
\alias{Sarim}
\docType{package}
\title{
\packageTitle{Sarim}
}
\description{
\packageDescription{Sarim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Sarim}
\packageIndices{Sarim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Sarim}

Maintainer: \packageMaintainer{Sarim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
