% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sarim_mcmc}
\alias{sarim_mcmc}
\title{MCMC sampler with Metropolis-Hasting step in Gibbs-sampler for use in sarim()-function}
\usage{
sarim_mcmc(y, eta_first, Z, K, gamma, ka_start, ka_values, solver, family, link,
  nIter, Ntrials = 1, m = 50L, thr = 1e-04, display_progress = TRUE,
  constraint = "No")
}
\arguments{
\item{y}{Response variable, given as a vector, use as.numeric() if error occur.}

\item{eta_first}{First calculation of eta = Z_1 * gamma_1 + ... + Z_p * gamma_p}

\item{Z}{Design matrices of the covariates, given as list with sparse matrix, 
use e.g. as(matrix, "dgCMatrix") from library(Matrix). Use for example the
useful sx()-function for smoothing.}

\item{K}{Structure/penalty matrices for the coefficients, given as list, 
also with sparse matrix. Can be choosen in sx()-function.}

\item{gamma}{List of coefficient, given as vector. Row-length need to be the 
same as the columns of Z. Per starting default from uniform distribution 
is sampled, but a specific starting value can be given, using the sx()-function
in the formular, e.g. y ~ sx(x1, gamma = c(rep(1, 5))).}

\item{solver}{List of the solvers ("rue" or "lanczos") for sampling from a gaussian distribution, i.e. gamma ~ N(eta, Q)
with Q as precision matrix. Can be choosen in sx()-function.}

\item{family}{Can currently only be "binomial" or "poisson"}

\item{link}{Link need to be chooses correspondingly to the family function. If
family = "binomial", please choose link = "logit" or if family = "poisson", 
please choose link = "log". Currently no other link function is present.}

\item{nIter}{Number of iterations for MCMC-algorithm}

\item{Ntrials}{Number of trails, only interesting for binomial distribution}

\item{m}{Number of maximal Lanczos-iterations}

\item{thr}{threshold when the Lanczos-algorithm should stop}

\item{kappa}{Start value for kappa, given as a list and double/float value.}

\item{kappa_values}{Coefficients for kappa, given as list within as vector c(kappa_a, kappa_b).}
}
\value{
Return a list of values:
"coef_results" = result of the estimated coefficient, output given as matrix;
"kappa_result" = result of the estimated kappa (precision) parameters, output given as vector
}
\description{
This generates samples for the coefficients using a Metropolis-Hasting step in
the Gibbs-sampling for the precision parameters and further assumed that the 
response variable is either binomial or poisson. So family = "binomial"/"poisson" with
link = "logit"/"log" is choosen respectively.
}
