% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx.R
\name{sx}
\alias{sx}
\title{Defining smooth in sarim formulae}
\usage{
sx(x = x, knots = 5, penalty = "rw2", degree = 3, solver = "rue",
  ka_start = 1, ka_a = 1, ka_b = 1e-04, linear_constraint = "FALSE",
  gamma = NULL, Z = NULL, K = NULL)
}
\arguments{
\item{x}{covariates to be smooth, similar to bs form spline package.}

\item{knots}{number of knots, only equidistant is acutual available.}

\item{penalty}{"identity", "rw1", "rw2" for structure/penalty matrix 
or "gmrf" if your own design matrix Z is given.}

\item{degree}{the order of the Basis functions, default = 3 (cubic splines).}

\item{solver}{"rue" or "lanczos" for sampling from a normal distribution. "rue" is
default and a good choose for normal smoothing. "lanczos" is an iterative methods
especially suitable for BIG STRUCTURE MATRICES, i.e. dim(Z) > 5000.}

\item{ka_start}{starting value for kappa .}

\item{ka_a}{a value for sampling form priori kappa~Ga(ka_a, ka_b), see kappa.}

\item{ka_b}{a value for sampling form priori kappa~Ga(ka_a, ka_b), see kappa.}

\item{linear_constraint}{apply a linear constraint if it needed.}

\item{gamma}{optional starting values for coefficent of gamma, the length(gamma)
should be the same as the columns of x, i.e. the number knots, which are  
choosen in this sx-function.}

\item{Z}{optional design matrix, given for spatial or temporal effects, in this case
the penalty need to be set to "gmrf". If Z is given to the function, an additional
structure/penalty matrix K need to be set up.}

\item{K}{structure/penalty matrix, given as the penalty for the prior parameters,
could be choose for example as intrinsic Gaussian Markov random field (IGMRF).}
}
\description{
Function used in definition of smooth terms within sarim model formulae.
}
